+++ __tests__/e2e/scenarios/create-slice-image-prompt.test.ts
@@ -13,7 +13,7 @@ test("should check slice creation (model, code, mocks) based on user image promp
 	)
 
 	const messages = await aiAgent.simulateUserQuery({
-		prompt: `Create a "Hero" slice based on my image [!image]: ${referenceScreenshotFile}`,
+		prompt: `Create and code a "Hero" slice based on my image [!image]: ${referenceScreenshotFile}`,
 	})
 	expect(messages.length).toBeGreaterThan(0)
 
@@ -23,10 +23,9 @@ test("should check slice creation (model, code, mocks) based on user image promp
 	expect(toolsUsed).toEqual(
 		expect.arrayContaining([
 			"how_to_model_slice",
-			"save_slice_model",
+			"save_slice_data",
 			"how_to_code_slice",
 			"how_to_mock_slice",
-			"verify_slice_mock",
 		]),
 	)
 
@@ -56,6 +55,7 @@ Focus on:
   -- structure is roughly the same
   -- text should be the same
 `,
+		testName: "create-slice-image-prompt_slice-creation-image-prompt",
 	})
 
 	console.info("Grade:", grade)