+++ src/lib/sentry.ts
@@ -25,9 +25,8 @@ type trackSentryErrorArgs = {
 		| "add_slice_to_custom_type"
 		| "how_to_code_slice"
 		| "how_to_model_slice"
-		| "save_slice_model"
+		| "save_slice_data"
 		| "how_to_mock_slice"
-		| "verify_slice_mock"
 	extra?: Record<string, unknown>
 }